+++ libs/sdk-core/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add `fieldData` to `Search.SmartResult`
 - Add `fields` in `SearchOptions`
+- Add `getResourceField` and `getResourceFieldBySlug` methods in `KnowledgeBox`
+- Add `ResourceFieldProperties` enum in kb models
 
 # 1.1.3 (2023-02-08)